vli: Rename the `vli_usbhub` plugin to `vli` and add RO support for Single PD
The protocol for com.vli.usbhub and com.vli.pd are 95% the same, so just split
out the common code into a subclass and move the GType into the quirk files.

In the future we'll be doing PD updates without any USB hub at all.
