+++ plugins/vli/fu-vli-usbhub-firmware.h
@@ -14,5 +14,5 @@
 G_DECLARE_FINAL_TYPE (FuVliUsbhubFirmware, fu_vli_usbhub_firmware, FU, VLI_USBHUB_FIRMWARE, FuFirmware)
 
 FuFirmware		*fu_vli_usbhub_firmware_new			(void);
-FuVliUsbhubDeviceKind	 fu_vli_usbhub_firmware_get_device_kind		(FuVliUsbhubFirmware	*self);
+FuVliDeviceKind		 fu_vli_usbhub_firmware_get_device_kind		(FuVliUsbhubFirmware	*self);
 guint16			 fu_vli_usbhub_firmware_get_device_id		(FuVliUsbhubFirmware	*self);
