+++ plugins/vli/fu-vli-usbhub-i2c-device.h
