+++ plugins/vli/meson.build
@@ -1,20 +1,23 @@
 cargs = ['-DG_LOG_DOMAIN="FuPluginVliUsbhub"']
 
 install_data([
-  'vli-usbhub.quirk',
+  'vli-pd.quirk',
   'vli-usbhub-lenovo.quirk',
+  'vli-usbhub.quirk',
   ],
   install_dir: join_paths(datadir, 'fwupd', 'quirks.d')
 )
 
-shared_module('fu_plugin_vli_usbhub',
+shared_module('fu_plugin_vli',
   fu_hash,
   sources : [
-    'fu-plugin-vli-usbhub.c',
+    'fu-plugin-vli.c',
+    'fu-vli-common.c',
+    'fu-vli-device.c',
+    'fu-vli-pd-device.c',
     'fu-vli-usbhub-common.c',
     'fu-vli-usbhub-device.c',
     'fu-vli-usbhub-firmware.c',
-    'fu-vli-usbhub-i2c-common.c',
     'fu-vli-usbhub-i2c-device.c',
     'fu-vli-usbhub-pd-common.c',
     'fu-vli-usbhub-pd-device.c',
